Trace IPv6 routes as well
The mechanism is fundamentally similar, but for some reason the ICMPv6
designers decided to use a completely different number scheme for
equivalent codes, making this more obnoxious than necessary.

I was unable to really test this as CenturyLink does not provide v6
routing.  I tried ::1 and fe80::1, and they seemed to behave reasonably.